Your commit message is incorrect based on the provided diff/code changes. Based on the diff you have given, the appropriate commit message should be:
Add calculatePi method to PiCalculator

Added a new static method 'calculatePi' to the PiCalculator utility class. This method calculates and returns the value of pi based on the number of terms provided, using a numerical series. This should enable dynamic and precise calculations of the Pi constant for various applications.