Add Hanoi class for Tower of Hanoi game
A new class named `Hanoi` is added to manage the Tower of Hanoi game. It includes static fields for three stacks as towers and a method called `solve` to solve the game. Also, a private method `move` utilized in `solve` to move discs between towers is included. This class lays the foundation for the Tower of Hanoi game functionalities.